코드리뷰 - 2차

See merge request cloud_track/class_01/web_project2/team03/pokeranger-back!16

+++ src/main/java/io/elice/pokeranger/user/service/UserService.java
@@ -7,6 +7,7 @@ import io.elice.pokeranger.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -16,8 +17,7 @@ public class UserService {
     private UserMapper userMapper;
 
     @Autowired
-    public UserService(UserRepository userRepository, UserMapper userMapper)
-    {
+    public UserService(UserRepository userRepository, UserMapper userMapper) {
         this.userRepository = userRepository;
         this.userMapper = userMapper;
     }
@@ -51,7 +51,15 @@ public class UserService {
     }
 
     public void deleteUser(Long userId) {
-        userRepository.deleteById(userId);
+        if (userRepository.existsById(userId)) {
+            userRepository.deleteById(userId);
+        } else {
+            throw new RuntimeException("User with ID " + userId + " not found");
+        }
+    }
+    public List<User> getAll() {
+
+        return userRepository.findAll();
     }
 
 }
